Refactor components to SearchDags folder.
